refactor: wine 도메인 엔드포인트 호출 함수 구조 변경 및 이름 재설정
