--- lib/api/wine/createWine.ts
@@ -1,9 +1,0 @@
-import { fetcher } from '@/lib/fetcher';
-import { CreateWineBody, CreateWineResponse } from './types';
-
-export function createWine(body: CreateWineBody) {
-  return fetcher<CreateWineResponse>('/api/proxy/wines', {
-    method: 'POST',
-    body,
-  });
-}
