--- lib/api/wine/getWines.ts
@@ -1,9 +1,0 @@
-import { fetcher } from '@/lib/fetcher';
-import { GetWinesQuery, GetWinesResponse } from './types';
-
-export function getWines(query: GetWinesQuery) {
-  return fetcher<GetWinesResponse>('/api/proxy/wines', {
-    method: 'GET',
-    query,
-  });
-}
