--- lib/api/wine/index.ts
@@ -1,7 +1,0 @@
-export * from './types';
-export * from './getWines';
-export * from './getWine';
-export * from './getRecommendedWines';
-export * from './createWine';
-export * from './updateWine';
-export * from './deleteWine';
