--- lib/api/wine/updateWine.ts
@@ -1,9 +1,0 @@
-import { fetcher } from '@/lib/fetcher';
-import { UpdateWineBody, UpdateWineResponse } from './types';
-
-export function updateWine(id: number, body: UpdateWineBody) {
-  return fetcher<UpdateWineResponse>(`/api/proxy/wines/${id}`, {
-    method: 'PATCH',
-    body,
-  });
-}
